[Move `@kbn/config-schema` to server] `@kbn/content-management-utils`
