+++ packages/kbn-content-management-utils/kibana.jsonc
@@ -1,5 +1,5 @@
 {
-  "type": "shared-common",
+  "type": "shared-server",
   "id": "@kbn/content-management-utils",
   "owner": "@elastic/kibana-data-discovery"
 }
