+++ src/plugins/dashboard/common/content_management/v2/types.ts
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-import {
+import type {
   ContentManagementCrudTypes,
   SavedObjectCreateOptions,
   SavedObjectUpdateOptions,
