+++ src/plugins/event_annotation/server/content_management/event_annotation_group_storage.ts
@@ -9,7 +9,7 @@ import { SOContentStorage } from '@kbn/content-management-utils';
 
 import { EVENT_ANNOTATION_GROUP_TYPE } from '@kbn/event-annotation-common';
 import { Logger } from '@kbn/logging';
-import { cmServicesDefinition } from '../../common/content_management/cm_services';
+import { cmServicesDefinition } from './schema/cm_services';
 import type { EventAnnotationGroupCrudTypes } from '../../common/content_management';
 
 export class EventAnnotationGroupStorage extends SOContentStorage<EventAnnotationGroupCrudTypes> {
