+++ src/plugins/links/server/content_management/links_storage.ts
@@ -10,7 +10,7 @@ import type { Logger } from '@kbn/logging';
 import { SOContentStorage } from '@kbn/content-management-utils';
 import { CONTENT_ID } from '../../common';
 import type { LinksCrudTypes } from '../../common/content_management';
-import { cmServicesDefinition } from '../../common/content_management/cm_services';
+import { cmServicesDefinition } from './schema/cm_services';
 
 export class LinksStorage extends SOContentStorage<LinksCrudTypes> {
   constructor({
