+++ src/plugins/saved_search/server/content_management/schema/v1/cm_services.ts
@@ -15,7 +15,10 @@ import {
   updateOptionsSchema,
   createResultSchema,
 } from '@kbn/content-management-utils';
-import { MIN_SAVED_SEARCH_SAMPLE_SIZE, MAX_SAVED_SEARCH_SAMPLE_SIZE } from '../../constants';
+import {
+  MIN_SAVED_SEARCH_SAMPLE_SIZE,
+  MAX_SAVED_SEARCH_SAMPLE_SIZE,
+} from '../../../../common/constants';
 
 const sortSchema = schema.arrayOf(schema.string(), { maxSize: 2 });
