+++ src/plugins/visualizations/common/content_management/v1/types.ts
@@ -18,7 +18,7 @@ import {
   CreateResult,
   UpdateResult,
 } from '@kbn/content-management-plugin/common';
-import { ContentManagementCrudTypes } from '@kbn/content-management-utils';
+import type { ContentManagementCrudTypes } from '@kbn/content-management-utils';
 
 import { VisualizationContentType } from '../types';
