+++ src/plugins/visualizations/public/vis_types/vis_type_alias_registry.ts
@@ -8,7 +8,7 @@
 
 import { SearchQuery } from '@kbn/content-management-plugin/common';
 import { ContentManagementPublicStart } from '@kbn/content-management-plugin/public';
-import {
+import type {
   ContentManagementCrudTypes,
   SavedObjectCreateOptions,
   SavedObjectUpdateOptions,
