+++ x-pack/plugins/maps/server/content_management/maps_storage.ts
@@ -9,7 +9,7 @@ import { SOContentStorage, tagsToFindOptions } from '@kbn/content-management-uti
 import { SavedObjectsFindOptions } from '@kbn/core-saved-objects-api-server';
 import type { Logger } from '@kbn/logging';
 import { CONTENT_ID } from '../../common/content_management';
-import { cmServicesDefinition } from '../../common/content_management/cm_services';
+import { cmServicesDefinition } from './schema/cm_services';
 import type { MapCrudTypes } from '../../common/content_management';
 
 const searchArgsToSOFindOptions = (args: MapCrudTypes['SearchIn']): SavedObjectsFindOptions => {
