+++ x-pack/plugins/maps/server/content_management/schema/v1/cm_services.ts
